+++ src/main/java/com/highgo/opendbt/teacher/domain/model/ExamForm.java
@@ -2,14 +2,14 @@ package com.highgo.opendbt.teacher.domain.model;
 
 public class ExamForm {
 
-	private int[] ids;
+	private Long[] ids;
 	private int[] values;
 
-	public int[] getIds() {
+	public Long[] getIds() {
 		return ids;
 	}
 
-	public void setIds(int[] ids) {
+	public void setIds(Long[] ids) {
 		this.ids = ids;
 	}
